requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
pymongo==4.6.0
motor==3.3.1
qdrant-client==1.6.9
openai==1.3.5
google-generativeai==0.3.1
pyyaml==6.0.1
numpy==1.26.2
tenacity==8.2.3
httpx==0.25.2 